cdp1802 testsuite: use long where appropriate on 16-bit targets
Avoids test failures due to extraneous warnings on 16-bit targets.

gcc/testsuite/

	* c-c++-common/Warray-bounds-3.c: Use long for 32-bit field on int16.
	* gcc.dg/analyzer/torture/switch-4.c: Use unsigned long for id field.
	* gcc.dg/cpp/ucs.c (foo): Use long for c.
	* gcc.dg/pr115646.c: Use long for bit fields wider than 16.
	* gcc.dg/pr93231.c: Use long for expression shifted more than 16 bits.

+++ gcc/testsuite/gcc.dg/pr93231.c
@@ -9,7 +9,7 @@ int ctz_ice1 (int x)
       31, 27, 13, 23, 21, 19, 16, 7, 26, 12, 18, 6, 11, 5, 10, 9
     };
 
-  return table[((int)((x & -x) * -0x077CB531)) >> 27];
+  return table[((long)((x & -x) * -0x077CB531)) >> 27];
 }
 
 int ctz_ice2 (unsigned x)
